+++ docs/CHANGELOG.rst
@@ -33,6 +33,7 @@ Added
 - Use generic permission filters when filtering variants by related models
 - Return only distinct ``VariantExperiment`` objects
 - Add ``--bam-output`` input argument to ``vc-gatk4-hc``
+- Add ``--max-mnp-distance`` input argument to ``vc-gatk4-hc``
 
 
 ===================